# api.yaml
openapi: 3.0.0
info:
  title: Hospital Management System API
  description: API documentation for the Hospital Management System V2. Includes routes for Authentication, Admin, Doctor, and Patient roles.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /doctor/availability:
    get:
      summary: Fetch slots submitted by the doctor
      responses:
        '200':
          description: Array of booked and open time slots
    post:
      summary: Upload new availability timeslots
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                slots:
                  type: array
                  items:
                    type: object
      responses:
        '201':
          description: Slots created
...
  /doctor/appointments/{id}/treatment:
    post:
      summary: Log treatment diagnosis after successful appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Treatment recorded, appointment closed
  /patient/doctors:
    get:
      summary: Search and find available doctors
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: specialization_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Filtered list of doctors and unbooked slots
  /patient/book:
    post:
      summary: Book an available timeslot
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                slot_id:
                  type: integer
      responses:
        '201':
          description: Appointment Confirmed
...